requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
